// src/values/enum.ts
export enum TYPE {
  TRAFFIC = "TRAFFIC",
  AFFECTION = "AFFECTION",
  CUISINE = "CUISINE",
  SPIRITUALITY = "SPIRITUALITY",
  SCHOOL = "SCHOOL",
  HEALTH = "HEALTH",
}

export enum NAME_TYPE {
  TRAFFIC = "Traffic",
  AFFECTION = "Affection",
  CUISINE = "Cuisine",
  SPIRITUALITY = "Spirituality",
  SCHOOL = "School",
  HEALTH = "Health",
}

export enum COLOR_TYPE {
  TRAFFIC = '#24A8AF',
  AFFECTION = '#DE3576',
  CUISINE = '#F5B120',
  SPIRITUALITY = '#5B6CA8',
  SCHOOL = '#24A8AF',
  HEALTH = '#DE3576',
}

export enum ICON_TYPE {
  TRAFFIC = 'road-variant',
  AFFECTION = 'cards-heart',
  CUISINE = 'food',
  SPIRITUALITY = 'eye-settings',
  SCHOOL = 'library',
  HEALTH = 'arm-flex',
}

export enum GENDER {
  MEN = 'MEN',
  WOMEN = 'WOMEN',
  CUSTOM = 'CUSTOM',
}